+++ src/TopicSubscribers.cpp
@@ -32,6 +32,11 @@ void TopicSubscribers<T>::subscribeCallBack(const typename T::ConstPtr& messages
     ss<<*messages;
     QString _data=QString::fromStdString(ss.str());
     QVariant new_data(_data);
+    
+    if (SignalsSlot::qstringList.rowCount() >= 100) {
+        SignalsSlot::qstringList.removeRows(0,SignalsSlot::qstringList.rowCount()-50);
+    }
+    
     SignalsSlot::qstringList.insertRows(SignalsSlot::qstringList.rowCount(), 1);
     SignalsSlot::qstringList.setData(SignalsSlot::qstringList.index(SignalsSlot::qstringList.rowCount()-1), new_data);
     Q_EMIT SignalsSlot::dataUpdated();