+++ src/YmlLoader.cpp
@@ -1,0 +1,115 @@
+//======================================================================*/
+//   Create Date: 2022/11/17
+//   ClassName  : YmlLoader
+//   Author     : mhjin julu1@naver.com
+//   Class Def  : yml file loader
+//======================================================================*/
+#include "YmlLoader.h"
+
+void YmlLoader::loadConfig( QString rvizFileName, rviz::Config *config) {
+    YAML::Node _rootNode=YAML::LoadFile(qPrintable(rvizFileName));
+    //printNodes(&_rootNode);
+    loadNodes(&_rootNode, config);
+}
+
+void YmlLoader::loadNodes(YAML::Node* node, rviz::Config* config) {
+    for(YAML::const_iterator it=node->begin(); it!=node->end(); it++) {
+        YAML::Node _node=it->first;
+        YAML::Node _value=it->second;
+        string nodeStr=_node.as<string>();
+        string valueStr;
+        rviz::Config childConfig;
+        switch ( _value.Type()) {
+            case YAML::NodeType::Scalar: 
+                valueStr=_value.as<string>();
+                //cout <<"NODE:"<< nodeStr << endl;
+                //cout <<"VALUE:"<< valueStr << endl;
+                config->mapSetValue(QString::fromStdString(nodeStr),QString::fromStdString(valueStr));
+                break;
+            case YAML::NodeType::Sequence:
+                //cout <<"NODE:"<< nodeStr <<"-SEQ-"<<_value.size()<< endl;
+                childConfig=config->mapMakeChild(QString::fromStdString(nodeStr));
+                
+                for (size_t i=0; i< _value.size(); i++) {
+                    YAML::Node arrayValue=_value[i];
+                    rviz::Config seqConfig=childConfig.listAppendNew();
+                    if (arrayValue.Type() == YAML::NodeType::Scalar) {
+                        valueStr=arrayValue.as<string>();
+                    //    cout<<"-" << valueStr<< endl;
+                        seqConfig.setValue(QString::fromStdString(valueStr));
+                    }
+                    else 
+                        loadNodes(&arrayValue, &seqConfig);
+                }
+                break;
+            case YAML::NodeType::Map:
+                //cout <<"NODE:"<< nodeStr <<"-map"<< endl;
+                childConfig=config->mapMakeChild(QString::fromStdString(nodeStr));
+                loadNodes(&_value, &childConfig);
+                break;
+            default :
+                cout <<"NONE:"<< _node.as<string>() << endl;
+        }
+    }
+}
+
+void YmlLoader::printNodes(YAML::Node* node) {
+    for(YAML::const_iterator it=node->begin(); it!=node->end(); it++) {
+        YAML::Node _node=it->first;
+        YAML::Node value=it->second;
+        cout<<_node.as<string>()<< ": ";
+        switch ( value.Type()) {
+            case YAML::NodeType::Scalar: 
+                cout << "(S)"<<value.as<string>()<<endl;
+                break;
+            case YAML::NodeType::Sequence:
+                cout <<"  - ";
+                for (size_t i=0; i< value.size(); i++) {
+                    YAML::Node arrayValue=value[i];
+                    if (arrayValue.Type() == YAML::NodeType::Scalar) {
+                        cout << "(A)"<<arrayValue.as<string>()<<endl;
+                    }
+                    else 
+                        printNodes(&arrayValue);
+                }
+                break;
+            case YAML::NodeType::Map:
+                printNodes(&value);
+                break;
+            default :
+                cout <<"NONE" << endl;
+        }
+    }
+}
+
+void YmlLoader::printConfig(rviz::Config* config) {
+    rviz::Config::MapIterator master_iter=config->mapIterator();;
+    for ( ; master_iter.isValid(); master_iter.advance()) {
+        QString key=master_iter.currentKey();
+        cout <<  qPrintable(key)<<" : ";
+        rviz::Config childConfig=master_iter.currentChild();
+        switch(childConfig.getType()) {
+            case rviz::Config::Type::Map:
+                printConfig(&childConfig);
+            break;
+            case rviz::Config::Type::List:
+                for (size_t i=0; i< childConfig.listLength(); i++) {
+                    rviz::Config listNode =childConfig.listChildAt(i);
+                    if (listNode.getType() == rviz::Config::Type::Value) {
+                        cout << "-" << qPrintable(listNode.getValue().toString())<<endl;
+                    }
+                    else 
+                        printConfig(&listNode);
+                }
+            break;
+            case rviz::Config::Type::Value:
+                cout << endl;
+                cout << qPrintable(childConfig.getValue().toString())<<endl;
+            break;
+            defaut:
+            break;
+
+        }
+    }
+}
+