+++ src/sideButtonActions.h
@@ -13,6 +13,9 @@
 #include "DAQViz.h"
 #include "RvizViewer.h"
 #include "RecordStarter.h"
+#include "ConfigLoader.h"
+#include "RosRunManager.h"
+#include "RecordPlayer.h"
 
 class SideButtonActions {
 public:
@@ -46,6 +49,9 @@ private:
     QMap <QString, QList<DAQViz*>* > *rvizMap;
     ros::MultiThreadedSpinner* spiner=NULL;
     RecordStarter recordStarter;
+    ConfigLoader config;
+    RosRunManager rosRunManager;
+    RecordPlayer recordPlayer;
 
 };
 #endif